Split the OS directory and calling of path
I have refactored the OS interaction of creating a directory and reduced the need to have multiple Consts being in main and the other packages into an internal package referred to as directory.go
